+++ Docs/api_docs/torch_visualization_quantization.rst
@@ -24,25 +24,26 @@ Code Examples
 
 **Required imports**
 
-.. literalinclude:: ../torch_code_examples/visualization.py
-    :language: python
-    :lines: 40, 42-43, 52-58
+.. literalinclude:: ../torch_code_examples/visualization_quantization.py
+   :language: python
+   :start-after: # Quantization visualization imports
+   :end-before: # End of import statements
 
 **Comparing Model After Optimization**
 
-.. literalinclude:: ../torch_code_examples/visualization.py
+.. literalinclude:: ../torch_code_examples/visualization_quantization.py
     :language: python
     :pyobject: visualize_changes_in_model_after_and_before_cle
 
 **Visualizing weight ranges in Model**
 
-.. literalinclude:: ../torch_code_examples/visualization.py
+.. literalinclude:: ../torch_code_examples/visualization_quantization.py
     :language: python
     :pyobject: visualize_weight_ranges_model
 
 **Visualizing Relative weight ranges in Model**
 
-.. literalinclude:: ../torch_code_examples/visualization.py
+.. literalinclude:: ../torch_code_examples/visualization_quantization.py
     :language: python
     :pyobject: visualize_relative_weight_ranges_model
